+++ src/Pointer.js
@@ -2,7 +2,7 @@
 // https://croquet.io
 // info@croquet.io
 
-import { GetPawn, RegisterMixin } from "@croquet/worldcore-kernel";
+import { GetPawn, RegisterMixin } from "./worldcore";
 
 //------------------------------------------------------------------------------------------
 //-- AM_PointerTarget ----------------------------------------------------------------------
@@ -639,6 +639,7 @@ export const PM_Pointer = superclass => class extends superclass {
         pe.id = wcEvent.id;
         pe.button = wcEvent.button;
         pe.buttons = wcEvent.buttons;
+        pe.instanceId = rc.instanceId;
         if (rc.ray) {
             pe.ray = {origin: rc.ray.origin.toArray(), direction: rc.ray.direction.toArray()};
         }