+++ src/dolbyChat.js
@@ -4,7 +4,7 @@
 
 /* global VoxeetSDK */
 
-import { Data, ViewService, v3_equals, q_equals, q_yaw } from "@croquet/worldcore-kernel";
+import { Data, ViewService, v3_equals, q_equals, q_yaw } from "./worldcore";
 
 let chatAudioMuted = false; // this state has to persist through dormancy
 let userSelectedMicLabel = null; // ditto
@@ -73,25 +73,27 @@ console.log(`DolbyChatManager (local actor ${alreadyHere ? "already" : "not yet"
     <div id='chatUI'>
         <div id='chatState' class='noselect'>
             <!-- <div id='worldName'></div> -->
-            <div id='chatCount' class='noselect'><span id='chatCountText'>0</span></div>
+            <div id='chatSymbol'></div>
+            <div id='peopleSymbol'></div>
+            <div id='chatCount' class='noselect'><span id='chatCountText'>-</span></div>
         </div>
         <div id='chatButtons'>
             <div id='toggleConnection' tabindex='4'>
-                <div class='buttonImage joined' title='leave voice chat'>Leave</div>
-                <div class='buttonImage notJoined' title='join voice chat'>Join</div>
+                <div class='buttonImage joined' title='leave voice chat'></div>
+                <div class='buttonImage notJoined' title='join voice chat'></div>
                 <div id='connection-tooltip' class='bouncing'>
                     <div id='connection-tooltip-arrow'></div>
                     <div id='connection-tooltip-contents' class='noselect'>join voice chat</div>
                 </div>
             </div>
             <div id='toggleAudio' tabindex='1'>
-                <div class='buttonImage enabled' title='mute mic'><i class="fa-solid fa-microphone"></i></div>
-                <div class='buttonImage disabled' title='unmute mic'><i class="fa-solid fa-microphone-slash"></i></div>
+                <div class='buttonImage enabled' title='mute mic'></div>
+                <div class='buttonImage disabled' title='unmute mic'></div>
                 <div class='buttonImage unavailable' title='mic unavailable'></div>
             </div>
             <div id='toggleSettings' tabindex='4'>
-                <div class='buttonImage enabled' title='hide settings'><i class="fa-solid fa-gear"></i></div>
-                <div class='buttonImage disabled' title='show settings'><i class="fa-solid fa-gear"></i></div>
+                <div class='buttonImage enabled' title='hide settings'></div>
+                <div class='buttonImage disabled' title='show settings'></div>
             </div>
         </div>
     </div>
@@ -114,7 +116,7 @@ console.log(`DolbyChatManager (local actor ${alreadyHere ? "already" : "not yet"
 </div>
 `;
             chatHolder = div.firstChild;
-            document.getElementById("world-info-container").append(chatHolder);
+            document.body.appendChild(chatHolder);
 
             ['toggleConnection', 'toggleAudio', 'toggleSettings', 'toggleMicrophoneTest'].forEach(buttonName => {
                 const elem = document.getElementById(buttonName);
@@ -145,11 +147,22 @@ console.log(`DolbyChatManager (local actor ${alreadyHere ? "already" : "not yet"
             loudnessValue: document.querySelector('#loudness .value'),
         };
 
+        this.uiStyles = {
+            preConnect: { width: '220px', height: '50px', transform: 'translate(-116px, 0px)' },
+            connected: { width: '220px', height: '50px', transform: 'translate(-116px, 0px)' },
+            settings: { width: '220px', height: '188px', transform: 'translate(-116px, 0px)' }
+        };
+        this.setUIStyle('preConnect');
+
         chatHolder.classList.remove('hidden');
     }
 
     setUIStyle(mode) {
-        Object.assign(this.elements.chatHolder.style);
+        const settings = this.uiStyles[mode];
+        const isNarrow = window.innerWidth < 600;
+        settings.left = isNarrow ? '60%' : '';
+        settings.top = isNarrow ? '60px' : '';
+        Object.assign(this.elements.chatHolder.style, settings);
     }
 
     toggleConnection() {
@@ -442,12 +455,7 @@ console.log(`DolbyChatManager (local actor ${alreadyHere ? "already" : "not yet"
         stream.getTracks().forEach(track => track.stop());
     }
 
-    stopAudioStream() {
-        if (this.measurableAudioStream) {
-            this.stopStream(this.measurableAudioStream);
-            delete this.measurableAudioStream;
-        }
-
+    stopMeasurableAudioStream() {
         if (this.measurableAudioStreamSource) {
             this.measurableAudioStreamSource.disconnect();
             delete this.measurableAudioStreamSource;
@@ -538,36 +546,34 @@ console.log(`DolbyChatManager (local actor ${alreadyHere ? "already" : "not yet"
 
         // how audio input works:
 
-        // we pass the selected device id to VoxeetSDK.  we also get the stream
-        // for the device, clone it, and from the clone make a mediaStreamSource
+        // we pass the selected device id to VoxeetSDK.  after that (because if
+        // we try *before*, at least on Chrome the stream provided to Voxeet will
+        // fail to support echo cancellation) we also get a stream for the device,
+        // and from it make a mediaStreamSource
         // (stored as this.measurableAudioStreamSource), which is connected to
         // the gainNode that was set up on initialisation.  the gainNode is
         // connected to a mediaStreamDestination node for local feedback testing
         // (this.testAudioNode), and to an analyser for measuring
         // local audio level.
 
-        // switching input device therefore involves
-        //   - requesting a stream from the specified device
-        //   - stopping the stream (if any) supplying local feedback
-        //   - making a mediaStreamSource from a clone of the new stream
-        //   - connecting the mediaStreamSource to the long-lived gainNode
+        console.log(`asking Voxeet to select device ID "${selectedId}"`);
+        const promise = this._setAudioInputPromise = VoxeetSDK.mediaDevice.selectAudioInput(selectedId)
+            .then(response => {
+                console.log("response from Voxeet:", response);
 
-        // jan 2021: avoid re-running getUserMedia on iPad, because there
-        // is only ever one audio input device, and a repeated getUserMedia
-        // causes the device to be somehow silenced (though not obviously
-        // muted, disabled, or ended).
-        let startPromise;
-        const isIPad = navigator.userAgent.match(/\biPad\b/);
-        const okToReplace = !isIPad || !this.chatAudioStream;
-        if (!okToReplace) {
-            console.log(`not invoking getUserMedia`);
-            startPromise = Promise.resolve(this.chatAudioStream);
-        } else {
-            console.log(`getUserMedia with device ID "${selectedId}"`);
-            startPromise = navigator.mediaDevices.getUserMedia({ audio: { deviceId: selectedId } });
-        }
-        const promise = this._setAudioInputPromise = startPromise
-            .then(stream => {
+                // jan 2021: avoid re-running getUserMedia on iPad - because there
+                // is only ever one audio input device, and a repeated getUserMedia
+                // causes the device to be somehow silenced (though not obviously
+                // muted, disabled, or ended).
+                const isIPad = navigator.userAgent.match(/\biPad\b/);
+                const okToReplace = !isIPad || !this.chatAudioStream;
+                if (!okToReplace) {
+                    console.log(`not invoking getUserMedia`);
+                    return this.chatAudioStream;
+                } else {
+                    return navigator.mediaDevices.getUserMedia({ audio: { deviceId: selectedId } });
+                }
+            }).then(stream => {
                 const chatAudioTrack = stream.getAudioTracks()[0];
                 const prevAudioTrack = this.chatAudioTrack;
                 if (!force && chatAudioTrack === prevAudioTrack) {
@@ -583,31 +589,14 @@ console.log(`DolbyChatManager (local actor ${alreadyHere ? "already" : "not yet"
                 chatAudioTrack.onunmute = () => {
                     console.log('audio track unmuted itself');
                 };
-                chatAudioTrack.onended = _event => {
-                    // if the track unexpectedly ends, it probably means
-                    // that something in the host's audio settings has
-                    // been changed or replaced.  force a refresh of
-                    // the audio input.
-                    console.warn("audio track ended");
-                    this.setAudioInput(true); // force re-init
-                };
 
-                // clone the stream (and its tracks) before using its track to
-                // create an Agora audio track.
-                const audioStreamClone = stream.clone();
-
-                // replace the cloned stream that feeds the
-                // level meter and the feedback test
-                this.stopAudioStream(); // also disconnects mediaStreamSource, if any
-                this.measurableAudioStream = audioStreamClone;
-                const mediaStreamSource = this.audioContext.createMediaStreamSource(audioStreamClone);
+                // replace the stream that feeds the level meter and the feedback
+                // test.
+                this.stopMeasurableAudioStream(); // disconnects mediaStreamSource, if any
+                const mediaStreamSource = this.audioContext.createMediaStreamSource(stream);
                 mediaStreamSource.connect(this.gainNode);
                 this.measurableAudioStreamSource = mediaStreamSource;
-                audioStreamClone.getAudioTracks()[0].onended = () => console.log(`local subsidiary audio track ended unexpectedly`);
 
-                VoxeetSDK.mediaDevice.selectAudioInput(selectedId).then(response => {
-                    console.log(response);
-                });
                 this.elements.toggleAudio.classList.remove('error');
                 this.mediaStarted = true;
             }).catch(err => {
@@ -650,7 +639,7 @@ console.log("unmuting local audio");
         const audioLevel = this.getLocalAudioLevel();
 
         // no need to display audio level if the meter isn't on view.
-        if (this.elements.chatHolder.classList.contains('hide-settings') || !this.measurableAudioStream) return;
+        if (this.elements.chatHolder.classList.contains('hide-settings') || !this.measurableAudioStreamSource) return;
 
         if (this._maxAudioLevelLongTerm === undefined || audioLevel > this._maxAudioLevelLongTerm) {
             this._maxAudioLevelLongTerm = audioLevel;